Only the needed code changes
